+++ src/melee/it/item.c
@@ -133,6 +133,17 @@ void HSD_JObjSetScaleItem(Item* it, HSD_JObj* jobj, Vec3* scl)
 }
 
 /// @private
+#ifdef MUST_MATCH
+inline
+#endif
+    void
+    HSD_JObjSetFacingDirItem(HSD_JObj* jobj, Item* it)
+{
+    if (it->xDC8_word.flags.x19 == true)
+        HSD_JObjSetRotationY(jobj, M_PI / 2 * it->facing_dir);
+}
+
+/// @private
 void func_80267130(HSD_GObj* item_gobj, SpawnItem* spawnItem)
 {
     Item* item_data = (Item*) HSD_GObjGetUserData(item_gobj);
